Clean up part 1
