+++ src/configuration.jl
@@ -1,16 +1,14 @@
-export AbstractSolution, Solution, NonSolution, Configuration, Configurations, solutions
-
 abstract type AbstractSolution end
 
 struct Solution <: AbstractSolution
-	x
+    x
 end
 
 struct NonSolution <: AbstractSolution
-	x
+    x
 end
 
-const Configuration{T} = T where T <: AbstractSolution # alias
+const Configuration{T} = T where T<:AbstractSolution # alias
 
 const Configurations{N} = Set{<:Configuration}
