+++ src/metrics.jl
@@ -1,5 +1,3 @@
-export hamming, minkowski, manhattan, weights_bias
-
 """
     hamming(x, X)
 Compute the hamming distance of `x` over a collection of solutions `X`, i.e. the minimal number of variables to switch in `x`to reach a solution.
